update and align with the ts starter
