+++ .eslintignore
@@ -2,4 +2,4 @@
 .git
 node_modules
 public
-build
+coverage
