--- .prettierrc.js
@@ -1,6 +1,0 @@
-module.exports = {
-  endOfLine: 'lf',
-  singleQuote: true,
-  trailingComma: 'all',
-  printWidth: 120,
-};
