--- .storybook/GlobalStyleDecorator.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-import { Theme } from '../src/components/ui/Theme';
-
-export const GlobalStyleDecorator = storyFn => <Theme>{storyFn()}</Theme>;
