--- .storybook/addons.js
@@ -1,2 +1,0 @@
-import '@storybook/addon-docs/register';
-import '@storybook/addon-actions/register';
