+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "awesome-gatsby-starter",
-  "version": "2.0.0",
+  "name": "@south-paw/awesome-gatsby-starter",
+  "version": "3.0.0",
   "description": "A starter for GatsbyJS with a preconfigured MDX, Storybook and ESLint environment",
   "keywords": [
     "gatsby",
@@ -28,46 +28,46 @@
     "serve": "gatsby serve"
   },
   "dependencies": {
-    "@mdx-js/mdx": "^1.5.1",
-    "@mdx-js/react": "^1.5.1",
-    "gatsby": "^2.17.11",
-    "gatsby-image": "^2.2.31",
-    "gatsby-plugin-manifest": "^2.2.27",
-    "gatsby-plugin-mdx": "^1.0.56",
-    "gatsby-plugin-offline": "^3.0.19",
-    "gatsby-plugin-react-helmet": "^3.1.14",
-    "gatsby-plugin-sharp": "^2.2.37",
-    "gatsby-plugin-styled-components": "^3.1.12",
-    "gatsby-source-filesystem": "^2.1.36",
-    "gatsby-transformer-sharp": "^2.3.3",
-    "polished": "^3.4.2",
+    "@mdx-js/mdx": "^1.6.6",
+    "@mdx-js/react": "^1.6.6",
+    "gatsby": "^2.23.4",
+    "gatsby-image": "^2.4.7",
+    "gatsby-plugin-manifest": "^2.4.11",
+    "gatsby-plugin-mdx": "^1.2.15",
+    "gatsby-plugin-offline": "^3.2.10",
+    "gatsby-plugin-react-helmet": "^3.3.4",
+    "gatsby-plugin-sharp": "^2.6.11",
+    "gatsby-plugin-styled-components": "^3.3.4",
+    "gatsby-source-filesystem": "^2.3.11",
+    "gatsby-transformer-sharp": "^2.5.5",
+    "polished": "^3.6.5",
     "prop-types": "^15.7.2",
-    "react": "^16.11.0",
-    "react-dom": "^16.11.0",
-    "react-helmet": "^5.2.1",
-    "react-uid": "^2.2.0",
-    "styled-components": "^4.4.1",
-    "styled-normalize": "^8.0.6"
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "react-helmet": "^6.1.0",
+    "react-uid": "^2.3.0",
+    "styled-components": "^5.1.1",
+    "styled-normalize": "^8.0.7"
   },
   "devDependencies": {
-    "@babel/core": "^7.7.2",
-    "@storybook/addon-actions": "^5.2.6",
-    "@storybook/addon-docs": "^5.2.6",
-    "@storybook/react": "^5.2.6",
-    "babel-eslint": "^10.0.3",
-    "babel-loader": "^8.0.6",
-    "cross-env": "^6.0.3",
-    "eslint": "^6.6.0",
-    "eslint-config-airbnb": "^18.0.1",
-    "eslint-config-prettier": "^6.5.0",
-    "eslint-plugin-compat": "^3.3.0",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-jest": "^23.0.3",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-prefer-object-spread": "^1.2.1",
-    "eslint-plugin-prettier": "^3.1.1",
-    "eslint-plugin-react": "^7.16.0",
-    "prettier": "^1.19.1",
-    "react-markdown": "^4.2.2"
+    "@babel/core": "^7.10.2",
+    "@storybook/addon-actions": "^6.0.0-beta.31",
+    "@storybook/addon-docs": "^6.0.0-beta.31",
+    "@storybook/addon-links": "^6.0.0-beta.31",
+    "@storybook/react": "^6.0.0-beta.31",
+    "babel-eslint": "^10.1.0",
+    "cross-env": "^7.0.2",
+    "eslint": "^7.2.0",
+    "eslint-config-airbnb": "^18.2.0",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-import": "^2.21.2",
+    "eslint-plugin-jest": "^23.13.2",
+    "eslint-plugin-jsx-a11y": "^6.3.0",
+    "eslint-plugin-prettier": "^3.1.4",
+    "eslint-plugin-react": "^7.20.0",
+    "eslint-plugin-react-hooks": "^4.0.4",
+    "prettier": "^2.0.5",
+    "react-is": "^16.13.1",
+    "rimraf": "^3.0.2"
   }
 }
