+++ src/components/Button/Button.stories.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import { Button } from '.';
+
+export default {
+  title: 'Button',
+  component: Button,
+};
+
+export const Simple = () => <Button>button</Button>;
