+++ src/components/Navigation/Navigation.stories.js
@@ -1,10 +1,9 @@
 import React from 'react';
-
 import { Navigation } from '.';
 
 export default {
-  title: 'Components|Navigation',
+  title: 'Navigation',
   component: Navigation,
 };
 
-export const simple = () => <Navigation />;
+export const Simple = () => <Navigation path="/" />;
