--- src/components/Site/styled.js
@@ -1,6 +1,0 @@
-import styled from 'styled-components';
-
-export const Main = styled.div`
-  margin: 32px 0;
-  min-height: calc(100vh - 56px - 64px - 97px);
-`;
