+++ src/components/Theme/index.js
@@ -1,10 +1,10 @@
-import React from 'react';
 import PropTypes from 'prop-types';
+import React from 'react';
 import { ThemeProvider } from 'styled-components';
 import { Normalize } from 'styled-normalize';
-
 import { BaseCSS } from './BaseCSS';
-import { system } from './system';
+
+const system = {};
 
 const Theme = ({ children }) => (
   <ThemeProvider theme={system}>
