--- src/components/ui/Button/index.js
@@ -1,1 +1,0 @@
-export * from './Button';
