--- src/components/ui/Container/index.js
@@ -1,1 +1,0 @@
-export * from './Container';
