--- src/components/ui/Theme/index.js
@@ -1,1 +1,0 @@
-export * from './Theme';
