--- src/components/ui/Theme/system.js
@@ -1,1 +1,0 @@
-export const system = {};
