+++ src/pages/markdown.mdx
@@ -2,7 +2,7 @@
 title: Markdown
 ---
 
-import { Button } from '../components/ui/Button';
+import { Button } from '../components/Button';
 
 # Markdown page example
